Migrate Vue filter to helper methods
https://v3-migration.vuejs.org/breaking-changes/filters.html
